Added lose condition
As of now, the only way for the player to fail and enter the game over state is not 'standing' on a platform, which is done via triggers. Because of this, have now added a Coroutine to move the player so they don't fall through the platforms.

Also, extending the platforms out to 9 units in length so the player can see what's next. May want to add some kind of fog/move the camera so they can't see the platforms spawn in.