Coins, Checker pattern, and changed Obstacle spawning
Updated the platforms to have a black/white checker pattern to make it easier to see them passing, along with seeing the foot obstacle. Have now added a coin object, and a way to spawn them. The Game Manager tracks the total coins and uses a scriptable object to save the total number of coins the player has between each try.